Fix model cleanup errors.
The model did not clean itself up correctly, which is turn lead to memory leaks
due to jobs never being ack'ed. This fixes it by picking out the right structure
in the data so you get the right acks going. It also asserts that we get the right
amount of acks whenever we clean up the system.